Rename getComputedStyle to computedStyle. Null check el/prop
