+++ src/js/utils/computed-style.js
@@ -1,3 +1,5 @@
+import window from 'global/window';
+
 /**
  * A safe getComputedStyle with an IE8 fallback.
  *
@@ -6,16 +8,20 @@
  * that the player doesn't break in these cases.
  * See https://bugzilla.mozilla.org/show_bug.cgi?id=548397 for more details.
  *
- * @function getComputedStyle
+ * @function computedStyle
  * @param el the element you want the computed style of
  * @param prop the property name you want
  */
-export default function getComputedStyle(el, prop) {
-  if (typeof getComputedStyle === 'function') {
-    const cs = getComputedStyle(el);
+export default function computedStyle(el, prop) {
+  if (!el || !prop) {
+    return '';
+  }
+
+  if (typeof window.getComputedStyle === 'function') {
+    const cs = window.getComputedStyle(el);
 
     return cs ? cs[prop] : '';
   }
 
-  return el.currentStyle[prop];
+  return el.currentStyle[prop] || '';
 }
